refactor: отредактировать параметры иконок и изменить структуру фалов слайдера
