+++ src/components/Dot/Dot.js
