+++ src/components/Dot/Dot.scss
@@ -1,0 +1,8 @@
+.dot {
+  background-image: url('./../Icon/dotNoFill.svg');
+  width: 12px;
+  height: 12px;
+  &_fill {
+    background-image: url('./../Icon/dotFill.svg');
+  }
+}
