+++ src/components/Icon/Icon_basket.jsx
@@ -11,20 +11,20 @@ const IconBasket = ({ className, ...props }) => {
         fill="none"
         xmlns="http://www.w3.org/2000/svg">
         <path
-          fill-rule="evenodd"
-          clip-rule="evenodd"
+          fillRule="evenodd"
+          clipRule="evenodd"
           d="M1.95976 0.480965C1.63833 -0.0118926 0.963198 -0.151079 0.481055 0.181064C-0.0118022 0.502492 -0.151119 1.1775 0.181024 1.65964L2.22762 4.73464L5.32393 16.3489C6.11679 19.3382 8.83813 21.4274 11.9346 21.4274H25.7132C26.3025 21.4274 26.7847 20.9453 26.7847 20.356C26.7847 19.7667 26.3025 19.2846 25.7132 19.2846H11.9346C10.2524 19.2846 8.73123 18.3846 7.89552 16.9917L26.0026 15.1489C27.4061 15.0096 28.5417 13.981 28.831 12.6096L29.9775 7.081C30.3954 5.08814 28.8633 3.21316 26.8276 3.21316H3.79186L1.95976 0.480965ZM7.15603 14.9132L4.60616 5.35602H26.8276C27.5133 5.35602 28.0168 5.97739 27.8776 6.64168L26.7311 12.1703C26.6346 12.631 26.2598 12.9739 25.7883 13.0168L7.15603 14.9132Z"
           fill="currentColor"
         />
         <path
-          fill-rule="evenodd"
-          clip-rule="evenodd"
+          fillRule="evenodd"
+          clipRule="evenodd"
           d="M6.42749 26.7846C6.42749 25.006 7.8632 23.5703 9.64178 23.5703C11.4203 23.5703 12.8561 25.006 12.8561 26.7846C12.8561 28.5632 11.4203 29.9989 9.64178 29.9989C7.8632 29.9989 6.42749 28.5632 6.42749 26.7846ZM9.64178 25.7132C9.05249 25.7132 8.57035 26.1953 8.57035 26.7846C8.57035 27.3739 9.05249 27.856 9.64178 27.856C10.2311 27.856 10.7132 27.3739 10.7132 26.7846C10.7132 26.1953 10.2311 25.7132 9.64178 25.7132Z"
           fill="currentColor"
         />
         <path
-          fill-rule="evenodd"
-          clip-rule="evenodd"
+          fillRule="evenodd"
+          clipRule="evenodd"
           d="M20.3562 26.7846C20.3562 25.006 21.7919 23.5703 23.5705 23.5703C25.3491 23.5703 26.7848 25.006 26.7848 26.7846C26.7848 28.5632 25.3491 29.9989 23.5705 29.9989C21.7919 29.9989 20.3562 28.5632 20.3562 26.7846ZM23.5705 25.7132C22.9812 25.7132 22.4991 26.1953 22.4991 26.7846C22.4991 27.3739 22.9812 27.856 23.5705 27.856C24.1598 27.856 24.6419 27.3739 24.6419 26.7846C24.6419 26.1953 24.1598 25.7132 23.5705 25.7132Z"
           fill="currentColor"
         />
