+++ src/components/Main/Main.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import './Main.scss';
+import ProductCard from '../ProductCard/ProductCard';
 
 const Main = () => {
   return <div className="main">
-    
+    <ProductCard/>
   </div>;
 };
