--- src/components/SliderImage/Dot.scss
@@ -1,8 +1,0 @@
-.dot {
-  background-image: url('./dotNoFill.svg');
-  width: 12px;
-  height: 12px;
-  &_fill {
-    background-image: url('./dotFill.svg');
-  }
-}
