--- src/components/SliderImage/images.js
@@ -1,7 +1,0 @@
-import img1 from './1.jpg';
-import img2 from './2.jpg';
-import img3 from './3.jpg';
-import img4 from './4.jpg';
-import img5 from './5.jpg';
-
-export const images = [img1, img2, img3, img4, img5];
